+++ packages/eip-5792-middleware/src/index.test.ts
@@ -1,0 +1,13 @@
+import * as allExports from '.';
+
+describe('@metamask/eip-5792-middleware', () => {
+  it('has expected JavaScript exports', () => {
+    expect(Object.keys(allExports)).toMatchInlineSnapshot(`
+      Array [
+        "processSendCalls",
+        "getCallsStatus",
+        "getCapabilities",
+      ]
+    `);
+  });
+});