+++ packages/eip-5792-middleware/src/index.ts
@@ -1,0 +1,11 @@
+export type {
+  ProcessSendCallsRequest,
+  ProcessSendCallsHooks,
+} from './methods/processSendCalls';
+export { processSendCalls } from './methods/processSendCalls';
+export { getCallsStatus } from './methods/getCallsStatus';
+export {
+  getCapabilities,
+  type GetCapabilitiesHooks,
+} from './methods/getCapabilities';
+export type { EIP5792Messenger } from './types';